Use AC_SEARCH_LIBS to find libutil
This demonstrates some of the problems I anticipate
in updating the build.  Validating the existence of
the file libutil.a or libutils.a is not much of a check,
and AC_SEARCH_LIBS is a more robust test.  However,
I can't tell what feature was needed from this
library, and the commit log is not much help,
since it only says: "Made a superb configure script."
AFAIK, libutil is an attempt to provide gnulib type
functionality at a link level, so I am going to
check for printf.