Add --with-socket-dir-remote option
This option allows the user to indicate that the SOCKDIR
is on a network mount.  It would perhaps be better to
use a special syntax on the file name to indicate this,
(eg '--with-socket-dir=remote:/path') or to actually
test the feature of the path, but this seems acceptable
and easier.