Do not append -YBSD to CC
CC is a user variable, and the maintainer
should not adjust it in the build.  See addition
to 'notes' for more.  This commit also includes
some replacement of raw 'if' with AS_IF and
cleans up some of the 'butterfly' checking.  Note
that checking for systems like butterfly is
a mistake, as we ought to be testing features,
but I'm not yet ready to commit to modifying
such tests.