Make SOCKDIR an AC_ARG_VAR instead of AC_ARG_WITH
Using AC_ARG_VAR will make it much easier for site
configurations.  A user can simply specify values of
the form SOCKDIR=/foo in a file and set CONFIG_SITE
to the path of that file, and all the site configuration
settings will be used in the configure script.  Note
that it is certainly possible to do this with AC_ARG_VAR
as well, by setting with_socket_dir=/foo in $CONFIG_SITE,
but it is unusual syntax for persons unfamiliar
with autoconf.  Also, it is more intuitive to use
configure SOCKDIR=/foo than configure --with-socket-dir=/foo.
However, the biggest reason is simply that --with-foo
is intended to be used to indicate that the package
foo should be an allowed dependency.  For example,
--with-gtk or --with-X, or that foo is a feature you
want compiled in, eg --with-sockets.  It is unusual
to use --with-foo to set a configurable setting, IMO.

Also, this commit provides more robust validation
of the value entered, so that an invalid entry will
generate a configure time error rather than a compile
time error.  (eg, SOCKDIR='invalid " string' would
have caused a compiler error, but is now caught during
configure.)