Move ALLOW_SYSSCREENRC from acconfig.h into configure.ac.
For consistency, I'm electing not to use --enable-sysscreenrc.  2 reasons:
consistency, and ease of use.  I find it is easier to put
ALLOW_SYSSCREENRC=no into a CONFIG_SITE file than to do
enable_sysscreenrc=no.  Hmmm.  That's the silliest thing
I've ever written...

Anyway, I've made a decision to use ARG_VAR, for better
or worse, and I need to stick with it.  It's 6 of one
half dozen of the other, and will be easy enough to
change in the future.