Moved the rev parsing into m4/parse-rev.m4
All of this rev parsing cruft is obsolete.  I'm keeing
it around for now since I don't want to modify the
source code any more than absolutely necessary, but
I need to get it out of configure.ac so I don't
need to keep looking at it.  This is purely a
psychological health preventative measure.  That
is, to keep me from going crazy.