Moving GIT_REV definition from make time to configure time.
Attempting to determine the git rev at make time is not
going to work from a tarball, and even with the stderr
redirect, we still see an error on FreeBSD.  Overall,
defining the git rev this way is a bad idea, and we should
be dumping it to a version file and distributing that with
the tarball, but that's for later.  However, there is
a bigger problem here.  Because we have configure.ac
in a sudbirectory, the path to the GIT_DIR changes
for a build from the tarball.  I had planned to wait
on moving configure.ac up a level until acconfig.h
was ready to be removed, but let's do it now instead.