Initial use of gnulib's getloadavg
This adds loadav2.c which is just the basic functionality
from loadav.c useing getloadavg from gnulib.  This allows
us to remove all of the cruft from configure.ac related
to getloadavg, but I am still struggling with the correct
way to invoke gnulib.  This seems to unconditionally
build all of the gnulib code and use it even on systems
where the functionality exists.