Add gnumakefile and maintainer-makefile from gnulib.
These are to enable generating tarballs with the git version
in the name.  Note that to get this functionality, we must
have annotated tags in the git repository.  screen 4.1.0
was released quite some time ago, but I'm making this
commit be tagged 4.1.0w so that at least one annotated
tag exists.