Simplify utmp checks in configury.
Note that the tests have not been changed in any way.
Here, we use m4_pushdef to reduce code duplication, but
do not modifying the tests significantly.  The check for
ut_host ought to be replaced with invocations of
AC_CHECK_MEMBER, but in the interests of getting
things done more quickly, such modifications
will be delayed until some other time.