Validate VARTYPES without default values.
With no default set, the screen_VARTYPE macro was not
doing much of anything.  This corrects that and ensures
that user entered values are valid.  Note that in testing
this, I notice that the build fails on OS X if usrlimit
is set.  This failure is evident in the current master
branch as well.